fix: crash at higher wandb versions
The dependency `wandb` was pinned to version `0.13.5` as it broke for higher versions (see issue #39).
I've updated `wandb` as the pinned version requires a `numpy` version below Numpy 2.
Apparently, something changed in the handling of the config parameter, so passing it as a `OmegaConf.DictConfig`
object is not longer possible. Calling `OmegaConf.to_container` on it converts it to a primitive that wandb
can handle.

This way `wandb` can be updated to the latest version.